Change number removal in normalizeauthors
When a unicode name is present, the transformation leads
to bogus results:

test <-  "Kerstin J\U3e34653cnicke <email@domain.org>"
sub(pattern="[[:digit:]]{1,2}",x=test,replacement="")
[1] "Kerstin J\U3e34653cnicke <email@domain.org>"

Interestingly. when the construction is applied to a vector that
contains a UTF-8 string, the first two letters of _all_ vector
elements are removed. Fix this with useBytes=TRUE.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/snatm@78 edb9625f-4e0d-4859-8d74-9fd3b1da38cb