Get rid of indexing overload in initiate.respond
The basic algorithm in initiate.respond is simple:
Iterate over all threads, identify initiator and
respondents for each, and tabulate the counts for
all participants in the threads. Unfortunately, the
implementation is obfuscated by the excessive use
of magic indices.

Rewrite the algorithm based on a data frame with
textual column labels.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/snatm@85 edb9625f-4e0d-4859-8d74-9fd3b1da38cb