Handle pathological cases in createedges
If the frequent keywords are in fact not so frequent, the
case that only one message fulfils the criteria can arise.
However, dim(forest) will be NULL in this case, which breaks
an assumption when the forest is converted to a matrix.

Deriving the column indices fails in this case, too, so move
the corresponding code to the beginning of the function.

For pathological revision ranges, it could happen that there is
only a single message in the forest from the very beginning. However,
it does not seem worth to take care of this (hypothetical) corner
case, so just bail out.

Signed-off-by: Wolfgang Mauerer <wolfgang.mauerer@siemens.com>

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/snatm@89 edb9625f-4e0d-4859-8d74-9fd3b1da38cb