Adjust conditionals to work for return vectors
In recent versions of R, some of snatm's conditionals cause an error
because the functions class() and is.na() return a vector instead of a
single boolean value for the given input.

Adjust these statements to iterate the return vector to find the desired
class attribute and to check for any missing values in the given vector,
respectively.

Signed-off-by: Nicole Hoess <nicole.hoess@oth-regensburg.de>